+++ CHANGELOG.md
@@ -10,7 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add end to end tests
-- index: Add file index journal via option `--journal`
+- Add file index journal
+- Speedup incemental database updates via file index journal
 
 ### Fixed